Switch sample quantity to weight
Rather than having the sample quantity be a count of individual seeds,
change it to a weight, specifically grams. This is easier to measure for
larger quantities, and easier to deal compare since smaller seeds can
have an astronomical number of seeds per gram.

Fixes: https://github.com/jonner/seedcollection/issues/18

+++ seedctl/src/output/rows.rs
@@ -47,7 +47,7 @@ pub struct SampleRowFull {
     source: String,
     date: String,
     #[tabled(display_with = "table_display_option")]
-    quantity: Option<i64>,
+    quantity: Option<f64>,
 }
 
 impl TryFrom<Sample> for SampleRowFull {
@@ -147,7 +147,7 @@ pub struct SampleRowDetails {
     #[tabled(rename = "Collection Date")]
     date: String,
     #[tabled(display_with = "table_display_option")]
-    quantity: Option<i64>,
+    quantity: Option<f64>,
     certainty: Certainty,
     #[tabled(
         display_with = "table_display_germination",
@@ -248,7 +248,7 @@ pub struct AllocationRowFull {
     source: String,
     date: String,
     #[tabled(display_with = "table_display_option")]
-    quantity: Option<i64>,
+    quantity: Option<f64>,
     #[tabled(display_with = "table_display_option")]
     notes: Option<String>,
 }
